# Bug 1352908 - Optimize certain filters in the template compiler