[% ... FILTER none %] is optimized away. There is no cost for using it.
[% ... FILTER html %] is optimized to a direct call to
Bugzilla::Util::html_quote(), saving a method call to $stash->filter.
[% ... FILTER uri %] also saves a call to $stash->filter.
[% ... FILTER null %] avoids a filter call and always results in ''

without patch:
```
spent 2.53s (1.64+890ms) within Bugzilla::Template::Context::filter which was called 381645 times, avg 7µs/call
```

with patch:
```
spent 470ms (186+285) within Bugzilla::Template::Context::filter which was called 37008 times, avg 13µs/call
```